+++ CHANGELOG.md
@@ -6,7 +6,12 @@ This project follows [Semantic Versioning](http://semver.org/).
 ### Changed
 - Updated angular to v12 and ng-zorro-antd to v12.
 
-## [30.1.5] 2022-04-22
+## [30.1.6] 2020-05-02
+### Fixed
+- Off-list string values from an open-choice field are now usable in an
+  expression for another open-choice field.
+
+## [30.1.5] 2020-04-22
 ### Fixed
 - Support for multiple values coming from intialExpression or calculateExpresion
   on fields that are of type string or coding.